Use single list of documentation pages.
Previously, we used a separate list of the doc pages for the menu (in
deltacloud.org/content/_submenu_documentation.haml) and for the documentation
Overview page (in docs/documentation.mdown).

This fix adds a single source located in deltacloud.org/lib/documentation.rb
which is used both for the menu and the overview page. It makes
adding/renaming/moving the doc pages easier.