Users don't have to care about Target Image IDs
The whole system with Target Image IDs is confusing. If the user builds an
image and forgets to save the ID, it's not trivial to get it back.

Since aeolus-image can push by Image IDs as well, let's use those instead.
They're much easier to get to (aeolus-image list --images) and it makes the
whole process more consistent (there's only one ID users need to care about end
to end).

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>